[FEAT/#74] MyViewingParty 연결하기

+++ app/src/main/java/umc/everyones/lck/domain/model/response/mypage/ParticipateViewingPartyMypageModel.kt
@@ -1,13 +1,11 @@
 package umc.everyones.lck.domain.model.response.mypage
 
-import umc.everyones.lck.data.dto.response.mypage.ParticipateViewingPartyMypageResponseDto
-
 data class ParticipateViewingPartyMypageModel(
     val viewingParties: List<ParticipateViewingPartyMypageElementModel>,
     val last: Boolean
 ) {
     data class ParticipateViewingPartyMypageElementModel(
-        val id: Int,
+        val id: Long,
         val name: String,
         val date: String
     )
